refactor: delete dead auth chat service shells
