--- backend/web/services/auth_service.py
@@ -1,5 +1,0 @@
-"""Compatibility shell for neutral auth service owner."""
-
-from backend.auth_service import AuthService
-
-__all__ = ["AuthService"]
