--- backend/web/services/chat_events.py
@@ -1,5 +1,0 @@
-"""Compatibility shell for messaging realtime events."""
-
-from messaging.realtime.events import ChatEventBus
-
-__all__ = ["ChatEventBus"]
