--- backend/web/services/contact_bootstrap_service.py
@@ -1,5 +1,0 @@
-"""Compatibility shell for neutral contact bootstrap owner."""
-
-from backend.contact_bootstrap import ensure_owner_agent_contact
-
-__all__ = ["ensure_owner_agent_contact"]
