--- backend/web/services/typing_tracker.py
@@ -1,5 +1,0 @@
-"""Compatibility shell for messaging typing tracker."""
-
-from messaging.realtime.typing import TypingTracker
-
-__all__ = ["TypingTracker"]
